+++ rust/stackablectl/CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added experimental `debug` command ([#204]).
 - Support listing endpoints of Listeners in `stackablectl stacklet list` command.
   Currently only HDFS is using listener-op, so we can only test that so far ([#213], [#219]).
 
@@ -21,6 +22,7 @@ All notable changes to this project will be documented in this file.
 
 [#181]: https://github.com/stackabletech/stackable-cockpit/pull/181
 [#202]: https://github.com/stackabletech/stackable-cockpit/pull/202
+[#204]: https://github.com/stackabletech/stackable-cockpit/pull/204
 [#213]: https://github.com/stackabletech/stackable-cockpit/pull/213
 
 ## [23.11.3] - 2024-01-03